chore: Convert some WebPlatformPlayer js file to ts file. Extract callback logic from WebPlatformPlayer index.
