--- apps/ledger-live-desktop/jsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "allowSyntheticDefaultImports": false,
-    "baseUrl": ".",
-    "paths": {
-      "~/*": ["./src/*"]
-    }
-  },
-  "exclude": ["node_modules", "dist"]
-}
