+++ apps/ledger-live-desktop/package.json
@@ -66,6 +66,9 @@
     "@sentry/tracing": "6.19.2",
     "@tippyjs/react": "^4.2.6",
     "@trust/keyto": "^1.0.1",
+    "@types/react": "^17.0.2",
+    "@types/react-redux": "^7.1.24",
+    "@types/styled-components": "^5.1.25",
     "@xstate/inspect": "^0.4.1",
     "@xstate/react": "^1.6.3",
     "animated": "^0.2.2",
