+++ libs/ledger-live-common/src/platform/JSONRPCServer.ts
@@ -4,15 +4,16 @@ import {
   JSONRPCServerAndClient,
   JSONRPCClient,
 } from "json-rpc-2.0";
-type LedgerLiveAPIHandlers = Record<string, () => any>;
+type LedgerLiveAPIHandlers = Record<string, (any) => any>;
 type SendFunc = (request: any) => Promise<void>;
 export const useJSONRPCServer = (
   handlers: LedgerLiveAPIHandlers,
   send: SendFunc
-) => {
+): Array<SendFunc> => {
   const serverRef: {
     current: null | JSONRPCServerAndClient;
   } = useRef(null);
+
   useEffect(() => {
     const server = new JSONRPCServerAndClient(
       new JSONRPCServer(),
@@ -29,5 +30,6 @@ export const useJSONRPCServer = (
       await serverRef.current.receiveAndSend(request);
     }
   }, []);
+
   return [receive];
 };
